requirements.txt
apscheduler>=3.10.4
feedparser>=6.0.11
langchain-core>=0.3.0
langchain-openai>=0.3.0
langgraph>=0.2.60
pydantic>=2.9.0
python-dotenv>=1.0.1
requests>=2.32.3